fix: adapt to new way of configuring link speed calculator
