+++ core/src/main/java/org/eqasim/core/components/traffic/EqasimTrafficQSimModule.java
@@ -2,7 +2,6 @@ package org.eqasim.core.components.traffic;
 
 import org.eqasim.core.components.config.EqasimConfigGroup;
 import org.matsim.core.mobsim.qsim.AbstractQSimModule;
-import org.matsim.core.mobsim.qsim.qnetsimengine.linkspeedcalculator.LinkSpeedCalculator;
 
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
@@ -10,12 +9,13 @@ import com.google.inject.Singleton;
 public class EqasimTrafficQSimModule extends AbstractQSimModule {
 	@Override
 	protected void configureQSim() {
-		bind(LinkSpeedCalculator.class).to(EqasimLinkSpeedCalculator.class);
+		addLinkSpeedCalculator().to(EqasimLinkSpeedCalculator.class);
+		bind(EqasimLinkSpeedCalculator.class).to(DefaultEqasimLinkSpeedCalculator.class);
 	}
 
 	@Provides
 	@Singleton
-	public EqasimLinkSpeedCalculator provideBaselineLinkSpeedCalculator(EqasimConfigGroup eqasimConfig) {
-		return new EqasimLinkSpeedCalculator(eqasimConfig.getCrossingPenalty());
+	public DefaultEqasimLinkSpeedCalculator provideDefaultEqasimLinkSpeedCalculator(EqasimConfigGroup eqasimConfig) {
+		return new DefaultEqasimLinkSpeedCalculator(eqasimConfig.getCrossingPenalty());
 	}
 }
