+++ core/src/main/java/org/eqasim/core/simulation/vdf/VDFQSimModule.java
@@ -1,13 +1,10 @@
 package org.eqasim.core.simulation.vdf;
 
+import org.eqasim.core.components.traffic.EqasimLinkSpeedCalculator;
 import org.eqasim.core.simulation.vdf.travel_time.VDFLinkSpeedCalculator;
 import org.eqasim.core.simulation.vdf.travel_time.VDFTravelTime;
-import org.matsim.api.core.v01.Scenario;
 import org.matsim.api.core.v01.population.Population;
-import org.matsim.core.api.experimental.events.EventsManager;
 import org.matsim.core.mobsim.qsim.AbstractQSimModule;
-import org.matsim.core.mobsim.qsim.qnetsimengine.ConfigurableQNetworkFactory;
-import org.matsim.core.mobsim.qsim.qnetsimengine.QNetworkFactory;
 
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
@@ -15,15 +12,7 @@ import com.google.inject.Singleton;
 public class VDFQSimModule extends AbstractQSimModule {
 	@Override
 	protected void configureQSim() {
-	}
-
-	@Provides
-	@Singleton
-	public QNetworkFactory provideQNetworkFactory(EventsManager events, Scenario scenario,
-			VDFLinkSpeedCalculator linkSpeedCalculator) {
-		ConfigurableQNetworkFactory networkFactory = new ConfigurableQNetworkFactory(events, scenario);
-		networkFactory.setLinkSpeedCalculator(linkSpeedCalculator);
-		return networkFactory;
+		bind(EqasimLinkSpeedCalculator.class).to(VDFLinkSpeedCalculator.class);
 	}
 
 	@Provides
