+++ core/src/main/java/org/eqasim/core/simulation/vdf/travel_time/VDFLinkSpeedCalculator.java
@@ -1,12 +1,12 @@
 package org.eqasim.core.simulation.vdf.travel_time;
 
+import org.eqasim.core.components.traffic.EqasimLinkSpeedCalculator;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.api.core.v01.population.Population;
 import org.matsim.core.mobsim.qsim.qnetsimengine.QVehicle;
-import org.matsim.core.mobsim.qsim.qnetsimengine.linkspeedcalculator.LinkSpeedCalculator;
 
-public class VDFLinkSpeedCalculator implements LinkSpeedCalculator {
+public class VDFLinkSpeedCalculator implements EqasimLinkSpeedCalculator {
 	private final VDFTravelTime travelTime;
 	private final Population population;
